requirements.txt
newspaper3k
kafka-python
psycopg2
